function2/requirements.txt
azure-servicebus
azure-cosmos
azure-identity
prettytable
pandas